refactor(davinci): split grown over-limit files hit by the allocator plumbing
